requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
numpy
torch
timm
trimesh
open3d
Pillow